change totalCost type to bigDecimal

+++ src/main/kotlin/com/trip_jr/tripJr/dto/booking/BookingDTO.kt
@@ -1,5 +1,6 @@
 package com.trip_jr.tripJr.dto.booking
 
+import java.math.BigDecimal
 import java.time.LocalDate
 import java.util.*
 
@@ -9,5 +10,5 @@ data class BookingDTO(
     val hotelId: UUID?,
     val checkInDate: LocalDate,
     val checkOutDate: LocalDate,
-    val totalCost: Double,
+    val totalCost: BigDecimal,
 )
